+++ Model/SubtaskTimeTrackingModelMod.php
@@ -259,11 +259,12 @@ class SubtaskTimeTrackingModelMod extends Base
      *
      * @access public
      * @param  integer   $task_id    Task id
+     * @param  bool      $use_ignores
      * @return bool
      */
-    public function updateTaskTimeTracking($task_id)
+    public function updateTaskTimeTracking($task_id, $use_ignores = true)
     {
-        $values = $this->calculateSubtaskTime($task_id);
+        $values = $this->calculateSubtaskTime($task_id, $use_ignores);
 
         return $this->db
                     ->table(\Kanboard\Model\TaskModel::TABLE)
@@ -276,11 +277,16 @@ class SubtaskTimeTrackingModelMod extends Base
      *
      * @access public
      * @param  integer   $task_id    Task id
+     * @param  bool      $use_ignores
      * @return array
      */
-    public function calculateSubtaskTime($task_id)
+    public function calculateSubtaskTime($task_id, $use_ignores = true)
     {
-        $ignore_subtask_titles = $this->helper->hoursViewHelper->getIgnoredSubtaskTitles();
+        if ($use_ignores) {
+            $ignore_subtask_titles = $this->helper->hoursViewHelper->getIgnoredSubtaskTitles();
+        } else {
+            $ignore_subtask_titles = [];
+        }
         return $this->db
                     ->table(\Kanboard\Model\SubtaskModel::TABLE)
                     ->eq('task_id', $task_id)